# sched : copy only the used experts when offloading prompt processing