When offloading prompt processing to GPU with MoE weights on CPU, all the experts are copied to VRAM. With this change, only the experts that are actually used are copied. With models where some experts are often unused, such as gpt-oss, this can reduce significantly the amount of data that needs to be copied to VRAM.

GGML_CUDA=1 scripts/compare-commits.sh master sl/sched-used-experts llama-bench -m gpt-oss-20b-mxfp4.gguf -ot "exps=CPU" -fa 1 -n 0 -p 2048 -ub "128-2048*2"

| Model                |   Microbatch size | Test   |   t/s master |   t/s sl/sched-used-experts |   Speedup |
|:---------------------|------------------:|:-------|-------------:|----------------------------:|----------:|
| gpt-oss 20B MXFP4 MoE |               128 | pp2048 |       124.24 |                      202.92 |      1.63 |
| gpt-oss 20B MXFP4 MoE |               256 | pp2048 |       253.18 |                      372.75 |      1.47 |
| gpt-oss 20B MXFP4 MoE |               512 | pp2048 |       502.73 |                      685.31 |      1.36 |
| gpt-oss 20B MXFP4 MoE |              1024 | pp2048 |      1015.20 |                     1178.50 |      1.16 |
| gpt-oss 20B MXFP4 MoE |              2048 | pp2048 |      1629.92 |                     2089.53 |      1.28 |



GGML_CUDA=1 scripts/compare-commits.sh master sl/sched-used-experts llama-bench -m gpt-oss-120b-mxfp4-00001-of-00003.gguf -ot "exps=CPU" -fa 1 -n 0 -p 2048 -ub "128-2048*2" -r 1

| Model                |   Microbatch size | Test   |   t/s master |   t/s sl/sched-used-experts |   Speedup |
|:---------------------|------------------:|:-------|-------------:|----------------------------:|----------:|
| gpt-oss 120B MXFP4 MoE |               128 | pp2048 |        22.37 |                       59.09 |      2.64 |
| gpt-oss 120B MXFP4 MoE |               256 | pp2048 |        41.80 |                       99.83 |      2.39 |
| gpt-oss 120B MXFP4 MoE |               512 | pp2048 |        87.93 |                      182.89 |      2.08 |
| gpt-oss 120B MXFP4 MoE |              1024 | pp2048 |       168.17 |                      255.84 |      1.52 |
| gpt-oss 120B MXFP4 MoE |              2048 | pp2048 |       345.01 |                      483.99 |      1.40 |

Short tests with real data and respecting the chat template (using `llama-cli -f <prompt.txt>`) also show a significant improvement.